+++ src/handlers/registry_handler.go
@@ -37,7 +37,7 @@ func GetEntities(w http.ResponseWriter, r *http.Request) {
 	// Retrieving name of node from query request
 	nodeName := r.URL.Query().Get("name")
 	confirmation, socket := checkNodeName(nodeName)
-	if confirmation == false {
+	if !confirmation  {
 
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
@@ -53,7 +53,7 @@ func GetEntities(w http.ResponseWriter, r *http.Request) {
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
 			Error: "Unexpected value found, height needs to be " +
-				"string of int!"})
+				"a string representing an int!"})
 		return
 	}
 
@@ -99,7 +99,7 @@ func GetNodes(w http.ResponseWriter, r *http.Request) {
 	// Retrieving name of node from query request
 	nodeName := r.URL.Query().Get("name")
 	confirmation, socket := checkNodeName(nodeName)
-	if confirmation == false {
+	if !confirmation  {
 
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
@@ -115,7 +115,7 @@ func GetNodes(w http.ResponseWriter, r *http.Request) {
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
 			Error: "Unexpected value found, height needs to be " +
-				"string of int!"})
+				"a string representing an int!"})
 		return
 	}
 
@@ -161,7 +161,7 @@ func GetRegistryEvents(w http.ResponseWriter, r *http.Request) {
 	// Retrieving name of node from query request
 	nodeName := r.URL.Query().Get("name")
 	confirmation, socket := checkNodeName(nodeName)
-	if confirmation == false {
+	if !confirmation  {
 
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
@@ -177,7 +177,7 @@ func GetRegistryEvents(w http.ResponseWriter, r *http.Request) {
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
 			Error: "Unexpected value found, height needs to be " +
-				"string of int!"})
+				"a string representing an int!"})
 		return
 	}
 
@@ -223,7 +223,7 @@ func GetRuntimes(w http.ResponseWriter, r *http.Request) {
 	// Retrieving name of node from query request
 	nodeName := r.URL.Query().Get("name")
 	confirmation, socket := checkNodeName(nodeName)
-	if confirmation == false {
+	if !confirmation  {
 
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
@@ -239,7 +239,7 @@ func GetRuntimes(w http.ResponseWriter, r *http.Request) {
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
 			Error: "Unexpected value found, height needs to be " +
-				"string of int!"})
+				"a string representing an int!"})
 		return
 	}
 
@@ -287,7 +287,7 @@ func GetRegistryStateToGenesis(w http.ResponseWriter, r *http.Request) {
 	// Retrieving name of node from query request
 	nodeName := r.URL.Query().Get("name")
 	confirmation, socket := checkNodeName(nodeName)
-	if confirmation == false {
+	if !confirmation  {
 
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
@@ -303,7 +303,7 @@ func GetRegistryStateToGenesis(w http.ResponseWriter, r *http.Request) {
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
 			Error: "Unexpected value found, height needs to be " +
-				"string of int!"})
+				"a string representing an int!"})
 		return
 	}
 
@@ -351,7 +351,7 @@ func GetEntity(w http.ResponseWriter, r *http.Request) {
 	// Retrieving name of node from query request
 	nodeName := r.URL.Query().Get("name")
 	confirmation, socket := checkNodeName(nodeName)
-	if confirmation == false {
+	if !confirmation  {
 
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
@@ -367,7 +367,7 @@ func GetEntity(w http.ResponseWriter, r *http.Request) {
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
 			Error: "Unexpected value found, height needs to be " +
-				"string of int!"})
+				"a string representing an int!"})
 		return
 	}
 
@@ -440,7 +440,7 @@ func GetNode(w http.ResponseWriter, r *http.Request) {
 	// Retrieving name of node from query request
 	nodeName := r.URL.Query().Get("name")
 	confirmation, socket := checkNodeName(nodeName)
-	if confirmation == false {
+	if !confirmation  {
 
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
@@ -456,7 +456,7 @@ func GetNode(w http.ResponseWriter, r *http.Request) {
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
 			Error: "Unexpected value found, height needs to be " +
-				"string of int!"})
+				"a string representing an int!"})
 		return
 	}
 
@@ -530,7 +530,7 @@ func GetNodeStatus(w http.ResponseWriter, r *http.Request) {
 	// Retrieving name of node from query request
 	nodeName := r.URL.Query().Get("name")
 	confirmation, socket := checkNodeName(nodeName)
-	if confirmation == false {
+	if !confirmation  {
 
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
@@ -546,7 +546,7 @@ func GetNodeStatus(w http.ResponseWriter, r *http.Request) {
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
 			Error: "Unexpected value found, height needs to be " +
-				"string of int!"})
+				"a string representing an int!"})
 		return
 	}
 
@@ -620,7 +620,7 @@ func GetRuntime(w http.ResponseWriter, r *http.Request) {
 	// Retrieving name of node from query request
 	nodeName := r.URL.Query().Get("name")
 	confirmation, socket := checkNodeName(nodeName)
-	if confirmation == false {
+	if !confirmation  {
 
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
@@ -636,7 +636,7 @@ func GetRuntime(w http.ResponseWriter, r *http.Request) {
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
 			Error: "Unexpected value found, height needs to be " +
-				"string of int!"})
+				"a string representing an int!"})
 		return
 	}