+++ src/handlers/scheduler_handler_test.go
@@ -50,7 +50,7 @@ func Test_GetValidators_InvalidHeight(t *testing.T) {
 			status, http.StatusOK)
 	}
 
-	expected := `{"error":"Unexpected value found, height needs to be string of int!"}`
+	expected := `{"error":"Unexpected value found, height needs to be a string representing an int!"}`
 
 	if strings.TrimSpace(rr.Body.String()) != strings.TrimSpace(expected) {
 		t.Errorf("handler returned unexpected body: got %v want %v",
@@ -130,7 +130,7 @@ func Test_GetCommittees_InvalidHeight(t *testing.T) {
 			status, http.StatusOK)
 	}
 
-	expected := `{"error":"Unexpected value found, height needs to be string of int!"}`
+	expected := `{"error":"Unexpected value found, height needs to be a string representing an int!"}`
 
 	if strings.TrimSpace(rr.Body.String()) != strings.TrimSpace(expected) {
 		t.Errorf("handler returned unexpected body: got %v want %v",
@@ -176,7 +176,7 @@ func Test_GetSchedulerStateToGenesis_InvalidHeight(t *testing.T) {
 			status, http.StatusOK)
 	}
 
-	expected := `{"error":"Unexpected value found, height needs to be string of int!"}`
+	expected := `{"error":"Unexpected value found, height needs to be a string representing an int!"}`
 
 	if strings.TrimSpace(rr.Body.String()) != strings.TrimSpace(expected) {
 		t.Errorf("handler returned unexpected body: got %v want %v",